+++ Assets/scripts/random_message.js
@@ -2,6 +2,7 @@ import { randomInt } from './utilities/math.js';
 
 var msg = document.getElementById('message');
 var new_message = "[MSG]";
+
 var messages = [
     "hope you enjoy your time here!",
     "you are in for a blast!",
@@ -10,5 +11,5 @@ var messages = [
     "Check out our projects if you want!"
 ];
 
-new_message = messages[randomInt(1, messages.length) - 1];
+new_message = messages[randomInt(0, messages.length - 1)];
 msg.innerHTML = new_message;