+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - v0.2.8.1 (duh)
 ### Added - Awsome new site
 - Home Page
+  - Random Message every time you come to the page (commits 1430 to 1434)
 - Project Page
 - Github Page
 - JSONs everywhere